Changes the way models are persisted and loaded
The algorithms are now defined by inheriting Algorithm class. This allows for more standardized persisting and loading. It also allows usage of multiple model binaries for prediction. Persisting and loading of this is automated through inheritance. So far only sklearn and keras serializing is supported, but can be easily extended for anything.